ppo 훈련 시 buffer sample 할 때 next observation을 double 자료형으로 가져오는 버그 수정

+++ link_rl/h_utils/buffers/buffer.py
@@ -253,7 +253,7 @@ class Buffer:
                 (
                     next_observations,
                     # torch.unsqueeze(torch.from_numpy(self.next_observation_value.numpy()).to(self.config.DEVICE), dim=0)
-                    torch.unsqueeze(self.next_observation_value, dim=0)
+                    torch.unsqueeze(self.next_observation_value, dim=0).float()
                 ),
                 dim=0
             )
